size_t != unsigned
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@6949 3db68df2-c116-0410-a063-a993310a9797

+++ libraries/fcollada/src/FCollada/FUtils/FUFileManager.cpp
@@ -389,7 +389,7 @@ fstring FUFileManager::GetApplicationFolderName()
 		}
 		else
 		{
-			path[max(1023u, size)] = '\0';
+			path[max(size_t(1023), size)] = '\0';
 			int i = stat (path, &stat_buf);
 			if (i == -1) break; 
 			else if (!S_ISLNK(stat_buf.st_mode)) break;
